Added/fixed a lot more stuff for s5 scripts.  May not be fully stable (s5 part)
git-svn-id: https://svn.code.sf.net/p/kaldi/code/trunk@903 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ egs/wsj/s5/steps/train_mono.sh
@@ -14,21 +14,23 @@ scale_opts="--transition-scale=1.0 --acoustic-scale=0.1 --self-loop-scale=0.1"
 num_iters=40    # Number of iterations of training
 max_iter_inc=30 # Last iter to increase #Gauss on.
 totgauss=1000 # Target #Gaussians.  
+boost_silence=1.0 # Factor by which to boost silence likelihoods in alignment
 realign_iters="1 2 3 4 5 6 7 8 9 10 12 14 16 18 20 23 26 29 32 35 38";
 config= # name of config file.
+stage=-4
 # End configuration section.
 
 if [ -f path.sh ]; then . ./path.sh; fi
 . parse_options.sh || exit 1;
 
 if [ $# != 3 ]; then
-   echo "Usage: steps/train_mono.sh [options] <data-dir> <lang-dir> <exp-dir>"
-   echo " e.g.: steps/train_mono.sh data/train.1k data/lang exp/mono"
-   echo "main options (for others, see top of script file)"
-   echo "  --config <config-file>                           # config containing options"
-   echo "  --nj <nj>                                        # number of parallel jobs"
-   echo "  --cmd (utils/run.pl|utils/queue.pl <queue opts>) # how to run jobs."
-   exit 1;
+  echo "Usage: steps/train_mono.sh [options] <data-dir> <lang-dir> <exp-dir>"
+  echo " e.g.: steps/train_mono.sh data/train.1k data/lang exp/mono"
+  echo "main options (for others, see top of script file)"
+  echo "  --config <config-file>                           # config containing options"
+  echo "  --nj <nj>                                        # number of parallel jobs"
+  echo "  --cmd (utils/run.pl|utils/queue.pl <queue opts>) # how to run jobs."
+  exit 1;
 fi
 
 data=$1
@@ -46,63 +48,70 @@ sdata=$data/split$nj;
 feats="ark,s,cs:apply-cmvn --norm-vars=false --utt2spk=ark:$sdata/JOB/utt2spk scp:$sdata/JOB/cmvn.scp scp:$sdata/JOB/feats.scp ark:- | add-deltas ark:- ark:- |"
 example_feats="`echo '$feats' | sed s/JOB/1/g`";
 
-echo "Initializing monophone system."
+echo "$0: Initializing monophone system."
 
-if [ -f $lang/phones/sets.int ]; then
-  shared_phones_opt="--shared-phones=$lang/phones/sets.int"
-fi
+[ ! -f $lang/phones/sets.int ] && exit 1;
+shared_phones_opt="--shared-phones=$lang/phones/sets.int"
 
+if [ $stage -le -3 ]; then
 # Note: JOB=1 just uses the 1st part of the features-- we only need a subset anyway.
-$cmd JOB=1 $dir/log/init.log \
- gmm-init-mono $shared_phones_opt "--train-feats=$feats subset-feats --n=10 ark:- ark:-|" $lang/topo 39 \
-   $dir/0.mdl $dir/tree || exit 1;
+  $cmd JOB=1 $dir/log/init.log \
+    gmm-init-mono $shared_phones_opt "--train-feats=$feats subset-feats --n=10 ark:- ark:-|" $lang/topo 39 \
+    $dir/0.mdl $dir/tree || exit 1;
+fi
 
 numgauss=`gmm-info --print-args=false $dir/0.mdl | grep gaussians | awk '{print $NF}'`
 incgauss=$[($totgauss-$numgauss)/$max_iter_inc] # per-iter increment for #Gauss
 
-rm $dir/.error 2>/dev/null
-
-echo "Compiling training graphs"
-$cmd JOB=1:$nj $dir/log/compile_graphs.JOB.log \
-  compile-train-graphs $dir/tree $dir/0.mdl  $lang/L.fst \
-   "ark:sym2int.pl --map-oov $oov_sym -f 2- $lang/words.txt < $sdata/JOB/text|" \
+if [ $stage -le -2 ]; then
+  echo "$0: Compiling training graphs"
+  $cmd JOB=1:$nj $dir/log/compile_graphs.JOB.log \
+    compile-train-graphs $dir/tree $dir/0.mdl  $lang/L.fst \
+    "ark:sym2int.pl --map-oov $oov_sym -f 2- $lang/words.txt < $sdata/JOB/text|" \
     "ark:|gzip -c >$dir/fsts.JOB.gz" || exit 1;
+fi
 
-echo "Aligning data equally (pass 0)"
-$cmd JOB=1:$nj $dir/log/align.0.JOB.log \
-  align-equal-compiled "ark:gunzip -c $dir/fsts.JOB.gz|" "$feats" ark,t:-  \| \
+if [ $stage -le -1 ]; then
+  echo "$0: Aligning data equally (pass 0)"
+  $cmd JOB=1:$nj $dir/log/align.0.JOB.log \
+    align-equal-compiled "ark:gunzip -c $dir/fsts.JOB.gz|" "$feats" ark,t:-  \| \
     gmm-acc-stats-ali --binary=true $dir/0.mdl "$feats" ark:- \
-      $dir/0.JOB.acc || exit 1;
-
+    $dir/0.JOB.acc || exit 1;
+fi
 
 # In the following steps, the --min-gaussian-occupancy=3 option is important, otherwise
 # we fail to est "rare" phones and later on, they never align properly.
 
-gmm-est --min-gaussian-occupancy=3  --mix-up=$numgauss \
-  $dir/0.mdl "gmm-sum-accs - $dir/0.*.acc|" $dir/1.mdl 2> $dir/log/update.0.log || exit 1;
+if [ $stage -le 0 ]; then
+  gmm-est --min-gaussian-occupancy=3  --mix-up=$numgauss \
+    $dir/0.mdl "gmm-sum-accs - $dir/0.*.acc|" $dir/1.mdl 2> $dir/log/update.0.log || exit 1;
+  rm $dir/0.*.acc
+fi
 
-rm $dir/0.*.acc
 
 beam=6 # will change to 10 below after 1st pass
 # note: using slightly wider beams for WSJ vs. RM.
 x=1
 while [ $x -lt $num_iters ]; do
-  echo "Pass $x"
-  if echo $realign_iters | grep -w $x >/dev/null; then
-    echo "Aligning data"
-    $cmd JOB=1:$nj $dir/log/align.$x.JOB.log \
-      gmm-align-compiled $scale_opts --beam=$beam --retry-beam=$[$beam*4] $dir/$x.mdl \
+  echo "$0: Pass $x"
+  if [ $stage -le $x ]; then
+    if echo $realign_iters | grep -w $x >/dev/null; then
+      echo "$0: Aligning data"
+      mdl="gmm-boost-silence --boost=$boost_silence `cat $lang/phones/optional_silence.csl` $dir/$x.mdl - |"
+      $cmd JOB=1:$nj $dir/log/align.$x.JOB.log \
+        gmm-align-compiled $scale_opts --beam=$beam --retry-beam=$[$beam*4] "$mdl" \
         "ark:gunzip -c $dir/fsts.JOB.gz|" "$feats" "ark,t:|gzip -c >$dir/ali.JOB.gz" \
-       || exit 1;
-  fi
-  $cmd JOB=1:$nj $dir/log/acc.$x.JOB.log \
-    gmm-acc-stats-ali  $dir/$x.mdl "$feats" "ark:gunzip -c $dir/ali.JOB.gz|" \
+        || exit 1;
+    fi
+    $cmd JOB=1:$nj $dir/log/acc.$x.JOB.log \
+      gmm-acc-stats-ali  $dir/$x.mdl "$feats" "ark:gunzip -c $dir/ali.JOB.gz|" \
       $dir/$x.JOB.acc || exit 1;
 
-  $cmd $dir/log/update.$x.log \
-    gmm-est --write-occs=$dir/$[$x+1].occs --mix-up=$numgauss $dir/$x.mdl \
+    $cmd $dir/log/update.$x.log \
+      gmm-est --write-occs=$dir/$[$x+1].occs --mix-up=$numgauss $dir/$x.mdl \
       "gmm-sum-accs - $dir/$x.*.acc|" $dir/$[$x+1].mdl || exit 1;
-  rm $dir/$x.mdl $dir/$x.*.acc $dir/$x.occs 2>/dev/null
+    rm $dir/$x.mdl $dir/$x.*.acc $dir/$x.occs 2>/dev/null
+  fi
   if [ $x -le $max_iter_inc ]; then
      numgauss=$[$numgauss+$incgauss];
   fi
@@ -112,11 +121,7 @@ done
 
 ( cd $dir; rm final.{mdl,occs} 2>/dev/null; ln -s $x.mdl final.mdl; ln -s $x.occs final.occs )
 
-# Print out summary of the warning messages.
-for x in $dir/log/*.log; do 
-  n=`grep WARNING $x | wc -l`; 
-  if [ $n -ne 0 ]; then echo $n warnings in $x; fi; 
-done
+utils/summarize_warnings.pl $dir/log
 
 echo Done
